\name{plot.lpExtPtr}
\alias{plot.lpExtPtr}

\title{lpExtPtr Plot Method}

\description{
Plots the feasible set of a simple linear program with two decision variables. The decision variables must be real, nonnegative and must not have a finite upper bound. Only inequality constraints are supported.
}

\usage{
\method{plot}{lpExtPtr}(x, y, ...)
}

\arguments{
  \item{x}{an lpSolve linear program model object.}
  \item{y}{this argument is ignored.}
  \item{\dots}{additional arguments are ignored.}
}

\author{Kjell Konis \email{kjell.konis@me.com}}

\keyword{hplot}


