\name{print.lpExtPtr}
\alias{print.lpExtPtr}

\title{lpSolve Print Method}

\description{
Display an lpSolve linear program model object in the console.
}

\usage{
\method{print}{lpExtPtr}(x, ...)
}

\arguments{
  \item{x}{an lpSolve linear program model object.}
  \item{\dots}{additional arguments are ignored.}
}

\value{
\code{x} is invisibly returned.
}

\references{\url{http://lpsolve.sourceforge.net/5.5/index.htm}}

\author{Kjell Konis \email{kjell.konis@me.com}}

\keyword{programming}


